Improve caching of `proto` in `GenerateProtoTask`
This change will remove the OS from the input properties of the
`GenerateProtoTask`. The expectation is that same versions of `protoc`
will produce same outputs regardless of the OS on which it was run.

If a protoc `artifact` is defined with GAV parameters, the classifier
and extension will not be considered part of the input. An exception is
made for snapshots, which will always be considered out of date and be
re-run.

If a protoc `path` is defined, then the path will be exposed as a File
Provider with absolute path normalization since there is no good way to
tell the version of `protoc` without running `protoc --version`.

See #457.